Second Commit
WORKING!
Notes on assignment:
Loop and Map.Entry Exercises

Write a program that does the following:
Asks the user for 5 numbers
Stores them in an array list
Finds the sum, product, largest, and smallest of those numbers

You are a car dealer. Create a hash map of vehicles.
The model is the Key, the make is the Value.
Ask the customer what car (model) they are looking for. Use the Hash Map to determine if you have that vehicle, and what make it is.
(e.g. "Oh, you're looking for a Civic? Our Honda selection is right over here...")